+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### next
 - Unreleased
+- Change: move the "no backend selected" compile error to the "-server" package instead of "-playback" (no need to specify a feature when compiling "termusic"(tui) now)
 - Fix: change default config ip address to `::1` instead of `::` (any old values on windows will need to be changed manually)
 
 ### [V0.9.1]